modified velodyne_clustering

+++ velodyne_clustering/include/velodyne_pointcloud/datacontainerbase.h
@@ -16,7 +16,7 @@ namespace velodyne_rawdata
         const float& y,
         const float& z,
         const uint16_t& ring,
-        const uint16_t& azimuth,
+        const float& azimuth,
         const float& distance,
         const float& intensity) = 0;
   };
